requirements.txt
discord.py==2.2.3
jishaku==2.5.1